Added trash page to store soft deleted records with delete option for perma delete and restore option for restoring the record

+++ app/Http/Service/FacultyServices/DeleteFacultyService.php
@@ -3,10 +3,21 @@ namespace App\Http\Service\FacultyServices;
 
 use App\Http\Service\ServicesInterface\DeleteFacultyServiceInterface;
 use App\Models\Faculty;
+use App\Models\Student;
 
 class DeleteFacultyService implements DeleteFacultyServiceInterface{
     public function deleteFaculty($rawData){
+        $studentsWithFaculty = Student::where('faculty', $rawData)->exists();
+
+        if ($studentsWithFaculty) {
+            throw new \Exception('Cannot delete faculty because there are students associated with it.');
+        }
         $faculty = Faculty::find($rawData);
-        $faculty->delete();
+
+        if ($faculty) {
+            $faculty->delete();
+        } else {
+            throw new \Exception('Faculty not found.');
+        }
     }
 }
